fix faces and center

+++ test/hypersphere.jl
@@ -1,5 +1,5 @@
 context("HyperSphere") do
-    @fact Sphere(Point(0,0,0), 4) --> HyperSphere{3,Int64}(Point{3,Int64}((0,0,0)),4)
+    @fact Sphere(Point(0,0,0), 4) --> HyperSphere{3,Int}(Point{3,Int}((0,0,0)),4)
     c = Circle(Point(0,0), 1)
     @fact c --> HyperSphere(Point(0,0), 1)
     @fact origin(c) --> Point(0,0)
